requirements.txt
connexion >= 2.6.0
connexion[swagger-ui] >= 2.6.0
flask_testing==0.8.1
python_dateutil == 2.6.0
requests==2.25.1
setuptools >= 21.0.0
swagger-ui-bundle >= 0.0.2

